==> 8_intersections.ipynb <==
import pandas as pd

TOPS = [50, 100, 150, 200, 250, 300, 350, 400, 450, 500]
FOLDER = './outputs/DS3'
MODELS = [
    'bertimbau-base',
    'bert-base-multilingual-cased',
    'bert-base-multilingual-uncased'
]

intersections = {}
dataset = ()
for top in TOPS:
    intersections[top] = {}
    for i in range(len(MODELS)):
        for j in range(i + 1, len(MODELS)):
            words_A = set(pd.read_csv(f'{FOLDER}/{MODELS[i]}/words/top_500_words.csv', sep=';').head(top).word.values)
            words_B = set(pd.read_csv(f'{FOLDER}/{MODELS[j]}/words/top_500_words.csv', sep=';').head(top).word.values)
            intersections[top][f'{MODELS[i]}_vs_{MODELS[j]}'] = len(words_A.intersection(words_B)) / top
pd.DataFrame(intersections)
